forgot password

+++ controllers/users_controller.js
@@ -27,9 +27,13 @@ module.exports.update = async function(req,res){
              
                 if (req.file){
 
-                    if (fs.existsSync(path.join(__dirname, '..', user.avatar))){
-                        fs.unlinkSync(path.join(__dirname, '..', user.avatar));
+                    if(user.avatar){
+                        if (fs.existsSync(path.join(__dirname, '..', user.avatar))){
+                            fs.unlinkSync(path.join(__dirname, '..', user.avatar));
+                        }
                     }
+
+                    
 
 
                     // this is saving the path of the uploaded file into the avatar field in the user
